==> notebooks/v2/Step6_Hyperparameter_Tuning.ipynb <==
# Iteration 1

import pickle

filename = r'..\..\data\v2\reg_gb_iteration1.pickle'
gb_model = pickle.load(open(filename,'rb'))

from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import GradientBoostingRegressor

gsc = GridSearchCV(
        estimator=GradientBoostingRegressor(),
        param_grid={
            'learning_rate': [0.1, 0.3, 0.5],
            'n_estimators': [100, 150, 200],
            'max_depth': [2, 3, 4, 5]
        },
        n_jobs=-1)

with open(r'..\..\data\v2\df_train_step3_subset1.pickle','rb') as flight_data_file:
     df_flights_TRAIN = pickle.load(flight_data_file)

with open(r'..\..\data\v2\df_test_step3_subset1.pickle','rb') as flight_data_file:
     df_flights_TEST = pickle.load(flight_data_file)

X_train = df_flights_TRAIN.drop('arr_delay',axis=1).to_numpy()
X_test = df_flights_TEST.drop('arr_delay',axis=1).to_numpy()
y_train = df_flights_TRAIN['arr_delay'].to_numpy()
y_test = df_flights_TEST['arr_delay'].to_numpy()

print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

grid_result = gsc.fit(X_train, y_train)
best_params = grid_result.best_params_

best_params

y_pred_train = grid_result.predict(X_train)
y_pred_test = grid_result.predict(X_test)

from sklearn.metrics import r2_score
r2_train = r2_score(y_train, y_pred_train)
r2_test = r2_score(y_test, y_pred_test)
print(r2_train)
print(r2_test)

import pandas as pd
import numpy as np
y_compare = pd.DataFrame({'y_test': y_test, 'y_pred_test': y_pred_test}, columns=['y_test', 'y_pred_test'])
y_compare

# Iteration 2

import pickle

filename = r'..\..\data\v2\reg_gb_iteration2.pickle'
gb_model = pickle.load(open(filename,'rb'))

from sklearn.model_selection import GridSearchCV
from sklearn.ensemble import GradientBoostingRegressor

gsc = GridSearchCV(
        estimator=GradientBoostingRegressor(),
        param_grid={
            'learning_rate': [0.09, 0.1, 0.11],
            'n_estimators': [100],
            'max_depth': [4]
        },
        n_jobs=-1)

with open(r'..\..\data\v2\df_train_step3_subset2.pickle','rb') as flight_data_file:
     df_flights_TRAIN = pickle.load(flight_data_file)

with open(r'..\..\data\v2\df_test_step3_subset2.pickle','rb') as flight_data_file:
     df_flights_TEST = pickle.load(flight_data_file)

X_train = df_flights_TRAIN.drop('arr_delay',axis=1).to_numpy()
X_test = df_flights_TEST.drop('arr_delay',axis=1).to_numpy()
y_train = df_flights_TRAIN['arr_delay'].to_numpy()
y_test = df_flights_TEST['arr_delay'].to_numpy()

print(X_train.shape)
print(X_test.shape)
print(y_train.shape)
print(y_test.shape)

grid_result = gsc.fit(X_train, y_train)
best_params = grid_result.best_params_

best_params

y_pred_train = grid_result.predict(X_train)
y_pred_test = grid_result.predict(X_test)

from sklearn.metrics import r2_score
r2_train = r2_score(y_train, y_pred_train)
r2_test = r2_score(y_test, y_pred_test)
print(r2_train)
print(r2_test)

import pandas as pd
import numpy as np
y_compare = pd.DataFrame({'y_test': y_test, 'y_pred_test': y_pred_test}, columns=['y_test', 'y_pred_test'])
y_compare

with open(r'..\..\data\v2\reg_gb_iteration2_tuned.pickle','wb') as flight_data_file:
     pickle.dump(grid_result, flight_data_file)
