==> notebooks/v2/Step7_Prediction&Deploymentv2.ipynb <==
# Based on Iteration 1

import pickle
import pandas as pd
import numpy as np

with open(r'..\..\data\flights_test.pickle','rb') as flight_data_file:
     df_test = pickle.load(flight_data_file)

df_test.shape

# drop features with more than 80% missing values

# df_test = df_test.dropna(thresh=df_test.shape[0]*0.8,axis=1)

# drop rows with null values

# df_test.dropna(inplace=True)

# drop rows with null taxi values

# df_test = df_test.replace({"taxi_out" : {0: np.nan},"taxi_in" : {0: np.nan}})

# df_test = df_test.reset_index(drop=True)

# fix datatypes

df_test['fl_date'] = pd.to_datetime(df_test['fl_date'])

df_test['crs_dep_time'] = df_test['crs_dep_time'].astype(int)
df_test['crs_dep_time'] = pd.to_datetime(df_test['crs_dep_time'], format = '%H%M', errors = 'coerce')
df_test = df_test.replace({"crs_dep_time" : {np.datetime64('NaT'): pd.to_datetime('2018-01-01 15:00:00')}})

# df_test['dep_time'] = df_test['dep_time'].astype(int)
# df_test['dep_time'] = pd.to_datetime(df_test['dep_time'], format = '%H%M', errors = 'coerce')

df_test['crs_arr_time'] = df_test['crs_arr_time'].astype(int)
df_test['crs_arr_time'] = pd.to_datetime(df_test['crs_arr_time'], format = '%H%M', errors = 'coerce')
df_test = df_test.replace({"crs_arr_time" : {np.datetime64('NaT'): pd.to_datetime('2018-01-01 15:00:00')}})

# df_test['arr_time'] = df_test['arr_time'].astype(int)
# df_test['arr_time'] = pd.to_datetime(df_test['arr_time'], format = '%H%M', errors = 'coerce')

# df_test['wheels_off'] = df_test['wheels_off'].astype(int)
# df_test['wheels_off'] = pd.to_datetime(df_test['wheels_off'], format = '%H%M', errors = 'coerce')

# df_test['wheels_on'] = df_test['wheels_on'].astype(int)
# df_test['wheels_on'] = pd.to_datetime(df_test['wheels_on'], format = '%H%M', errors = 'coerce')

df_test['mkt_carrier_fl_num'] = df_test['mkt_carrier_fl_num'].astype(str)
df_test['op_carrier_fl_num'] = df_test['op_carrier_fl_num'].astype(str)
df_test['origin_airport_id'] = df_test['origin_airport_id'].astype(str)
df_test['dest_airport_id'] = df_test['dest_airport_id'].astype(str)

df_test.info()

# drop datetime values that didn't convert

# df_test.dropna(inplace=True)
# df_test = df_test.reset_index(drop=True)

df_test.info()

df_test.head()

df_test['year'] = df_test['fl_date'].dt.year
df_test['month'] = df_test['fl_date'].dt.month
df_test['day'] = df_test['fl_date'].dt.day
df_test['day_of_week'] = df_test['fl_date'].dt.day_name()
df_test["is_weekend"] = df_test['fl_date'].dt.dayofweek > 4

with open(r'..\..\data\v2\season_dict.pickle','rb') as flight_data_file:
     season_dict = pickle.load(flight_data_file)

df_test["season"] = df_test["month"].map(season_dict)

from pandas.tseries.holiday import USFederalHolidayCalendar as calendar

dr = pd.date_range(start='2018-01-01', end='2019-12-31')

cal = calendar()
holidays = cal.holidays(start=dr.min(), end=dr.max())

df_test['is_holiday'] = df_test['fl_date'].isin(holidays)

df_test['crs_dep_time_hour'] = df_test['crs_dep_time'].dt.hour
df_test['crs_arr_time_hour'] = df_test['crs_arr_time'].dt.hour

cut_labels_5 = ['early morning', 'late morning', 'midday', 'late afternoon', 'evening']
cut_bins = [0, 5, 10, 15, 20, 24]
df_test['crs_dep_time_of_day'] = pd.cut(df_test['crs_dep_time_hour'], bins=cut_bins, labels=cut_labels_5)

cut_labels_5 = ['early morning', 'late morning', 'midday', 'late afternoon', 'evening']
cut_bins = [0, 5, 10, 15, 20, 24]
df_test['crs_arr_time_of_day'] = pd.cut(df_test['crs_arr_time_hour'], bins=cut_bins, labels=cut_labels_5)

new_origin = df_test['origin_city_name'].str.split(",", n = 1, expand = True)
new_dest = df_test['dest_city_name'].str.split(",", n = 1, expand = True)
df_test['origin_state'] = new_origin[1].str.strip()
df_test['dest_state'] = new_dest[1].str.strip()

with open(r'..\..\data\v2\state_dict.pickle','rb') as flight_data_file:
     state_dict = pickle.load(flight_data_file)

df_test["origin_region"] = df_test["origin_state"].map(state_dict)
df_test["dest_region"] = df_test["dest_state"].map(state_dict)

# df_test['mean_mkt_carrier_delay'] = df_test.groupby(['mkt_unique_carrier'])['arr_delay'].transform('mean')
# df_test['mean_op_carrier_delay'] = df_test.groupby(['op_unique_carrier'])['arr_delay'].transform('mean')

# df_test['mean_origin_airport_delay'] = df_test.groupby(['origin_airport_id'])['arr_delay'].transform('mean')
# df_test['mean_dest_airport_delay'] = df_test.groupby(['dest_airport_id'])['arr_delay'].transform('mean')

# df_test['mean_flight_delay'] = df_test.groupby(['op_unique_carrier','op_carrier_fl_num','crs_dep_time_of_day','origin_city_name','dest_city_name'])['arr_delay'].transform('mean')

df_test.head()

df_test = df_test[['day_of_week', 'is_weekend',
       'season', 'is_holiday', 'crs_dep_time_of_day', 'crs_arr_time_of_day', 'origin_region', 'dest_region']]

nom_feats_test = df_test.dtypes[(df_test.dtypes == 'object') | (df_test.dtypes == 'category') ].index.tolist()
bool_feats_test = df_test.dtypes[df_test.dtypes == 'bool'].index.tolist()

cat_feats_test = nom_feats_test + bool_feats_test

df_numeric_test = df_test.drop(columns=cat_feats_test)

df_dummy_test = pd.get_dummies(df_test[nom_feats_test])
df_dummy_test.head() #drop_first=True?

df_dummy_test['season_Spring'] = 0
df_dummy_test['season_Fall'] = 0
df_dummy_test['season_Summer'] = 0

df_test.replace({
    "is_weekend" : {True : 1, False : 0}, 
    "is_holiday" : {True : 1, False : 0},
    }, inplace=True)

transformed_df_test = df_dummy_test.merge(df_test[["is_weekend","is_holiday"]], left_index=True, right_index=True)

with open(r'..\..\data\v2\reg_gb_iteration1.pickle','rb') as flight_data_file:
     reg = pickle.load(flight_data_file)

y_pred = reg.predict(transformed_df_test)

y_pred

with open(r'..\..\data\flights_test.pickle','rb') as flight_data_file:
     df_test = pickle.load(flight_data_file)

submission = df_test[['fl_date', 'mkt_carrier',
       'mkt_carrier_fl_num', 'origin', 'dest']]

submission.head()

import pandas as pd

# submission = submission.dropna()
submission.info()

submission['predicted_delay'] = y_pred

submission.head()

submission.to_csv(r'..\..\data\submission.csv', index_label=False)

# Based on Iteration 2

import pickle
import pandas as pd
import numpy as np

with open(r'..\..\data\flights_test.pickle','rb') as flight_data_file:
     df_test = pickle.load(flight_data_file)

df_test.shape

# fix datatypes

df_test['fl_date'] = pd.to_datetime(df_test['fl_date'])

df_test['crs_dep_time'] = df_test['crs_dep_time'].astype(int)
df_test['crs_dep_time'] = pd.to_datetime(df_test['crs_dep_time'], format = '%H%M', errors = 'coerce')
df_test = df_test.replace({"crs_dep_time" : {np.datetime64('NaT'): pd.to_datetime('2018-01-01 15:00:00')}})

# df_test['dep_time'] = df_test['dep_time'].astype(int)
# df_test['dep_time'] = pd.to_datetime(df_test['dep_time'], format = '%H%M', errors = 'coerce')

df_test['crs_arr_time'] = df_test['crs_arr_time'].astype(int)
df_test['crs_arr_time'] = pd.to_datetime(df_test['crs_arr_time'], format = '%H%M', errors = 'coerce')
df_test = df_test.replace({"crs_arr_time" : {np.datetime64('NaT'): pd.to_datetime('2018-01-01 15:00:00')}})

# df_test['arr_time'] = df_test['arr_time'].astype(int)
# df_test['arr_time'] = pd.to_datetime(df_test['arr_time'], format = '%H%M', errors = 'coerce')

# df_test['wheels_off'] = df_test['wheels_off'].astype(int)
# df_test['wheels_off'] = pd.to_datetime(df_test['wheels_off'], format = '%H%M', errors = 'coerce')

# df_test['wheels_on'] = df_test['wheels_on'].astype(int)
# df_test['wheels_on'] = pd.to_datetime(df_test['wheels_on'], format = '%H%M', errors = 'coerce')

df_test['mkt_carrier_fl_num'] = df_test['mkt_carrier_fl_num'].astype(str)
df_test['op_carrier_fl_num'] = df_test['op_carrier_fl_num'].astype(str)
df_test['origin_airport_id'] = df_test['origin_airport_id'].astype(str)
df_test['dest_airport_id'] = df_test['dest_airport_id'].astype(str)

df_test.info()

df_test.head()

df_test['year'] = df_test['fl_date'].dt.year
df_test['month'] = df_test['fl_date'].dt.month
df_test['day'] = df_test['fl_date'].dt.day
df_test['day_of_week'] = df_test['fl_date'].dt.day_name()
df_test["is_weekend"] = df_test['fl_date'].dt.dayofweek > 4

with open(r'..\..\data\v2\season_dict.pickle','rb') as flight_data_file:
     season_dict = pickle.load(flight_data_file)

df_test["season"] = df_test["month"].map(season_dict)

from pandas.tseries.holiday import USFederalHolidayCalendar as calendar

dr = pd.date_range(start='2018-01-01', end='2019-12-31')

cal = calendar()
holidays = cal.holidays(start=dr.min(), end=dr.max())

df_test['is_holiday'] = df_test['fl_date'].isin(holidays)

df_test['crs_dep_time_hour'] = df_test['crs_dep_time'].dt.hour
df_test['crs_arr_time_hour'] = df_test['crs_arr_time'].dt.hour

cut_labels_5 = ['early morning', 'late morning', 'midday', 'late afternoon', 'evening']
cut_bins = [0, 5, 10, 15, 20, 24]
df_test['crs_dep_time_of_day'] = pd.cut(df_test['crs_dep_time_hour'], bins=cut_bins, labels=cut_labels_5)

cut_labels_5 = ['early morning', 'late morning', 'midday', 'late afternoon', 'evening']
cut_bins = [0, 5, 10, 15, 20, 24]
df_test['crs_arr_time_of_day'] = pd.cut(df_test['crs_arr_time_hour'], bins=cut_bins, labels=cut_labels_5)

new_origin = df_test['origin_city_name'].str.split(",", n = 1, expand = True)
new_dest = df_test['dest_city_name'].str.split(",", n = 1, expand = True)
df_test['origin_state'] = new_origin[1].str.strip()
df_test['dest_state'] = new_dest[1].str.strip()

with open(r'..\..\data\v2\state_dict.pickle','rb') as flight_data_file:
     state_dict = pickle.load(flight_data_file)

df_test["origin_region"] = df_test["origin_state"].map(state_dict)
df_test["dest_region"] = df_test["dest_state"].map(state_dict)

with open(r'..\..\data\v2\mkt_carrier_delay_dict.pickle','rb') as flight_data_file:
     mkt_carrier_delay_dict = pickle.load(flight_data_file)

df_test['mean_mkt_carrier_delay'] = df_test['mkt_unique_carrier'].map(mkt_carrier_delay_dict[0])
df_test['mean_mkt_carrier_delay'] = df_test['mean_mkt_carrier_delay'].fillna(5.115402956604673) # overall mean delay

with open(r'..\..\data\v2\op_carrier_delay_dict.pickle','rb') as flight_data_file:
     op_carrier_delay_dict = pickle.load(flight_data_file)

df_test['mean_op_carrier_delay'] = df_test['op_unique_carrier'].map(op_carrier_delay_dict[0])
df_test['mean_op_carrier_delay'] = df_test['mean_op_carrier_delay'].fillna(5.115402956604673) # overall mean delay

df_test.head()

df_test = df_test[['day_of_week', 'is_weekend',
       'season', 'is_holiday', 'crs_dep_time_of_day', 'crs_arr_time_of_day', 'origin_region', 'dest_region','mean_mkt_carrier_delay','mean_op_carrier_delay']]

nom_feats_test = df_test.dtypes[(df_test.dtypes == 'object') | (df_test.dtypes == 'category') ].index.tolist()
bool_feats_test = df_test.dtypes[df_test.dtypes == 'bool'].index.tolist()

cat_feats_test = nom_feats_test + bool_feats_test

df_numeric_test = df_test.drop(columns=cat_feats_test)

df_dummy_test = pd.get_dummies(df_test[nom_feats_test])
df_dummy_test.head() #drop_first=True?

df_dummy_test['season_Spring'] = 0
df_dummy_test['season_Fall'] = 0
df_dummy_test['season_Summer'] = 0

df_test.replace({
    "is_weekend" : {True : 1, False : 0}, 
    "is_holiday" : {True : 1, False : 0},
    }, inplace=True)

with open(r'..\..\data\v2\df_test_step3_subset2_scaler.pickle','rb') as flight_data_file:
     scaler = pickle.load(flight_data_file)

scaled_data_test = pd.DataFrame(scaler.transform(df_numeric_test), columns=df_numeric_test.columns)

transformed_df_test = scaled_data_test.merge(df_dummy_test, left_index=True, right_index=True)
transformed_df_test = transformed_df_test.merge(df_test[bool_feats_test], left_index=True, right_index=True)
# transformed_df_test = df_dummy_test.merge(df_test[["is_weekend","is_holiday"]], left_index=True, right_index=True)

transformed_df_test.head()

with open(r'..\..\data\v2\reg_gb_iteration2.pickle','rb') as flight_data_file:
     reg = pickle.load(flight_data_file)

y_pred = reg.predict(transformed_df_test)

y_pred

with open(r'..\..\data\flights_test.pickle','rb') as flight_data_file:
     df_test = pickle.load(flight_data_file)

submission = df_test[['fl_date', 'mkt_carrier',
       'mkt_carrier_fl_num', 'origin', 'dest']]

submission.head()

import pandas as pd

# submission = submission.dropna()
submission.info()

submission['predicted_delay'] = y_pred

submission.head()

submission.to_csv(r'..\..\data\submission.csv', index_label=False)
